Conflicts:
	docs/source/cython_wrapper.rst
	docs/source/getting_started.rst
	quantlib/_currency.pxd
	quantlib/_index.pxd
	quantlib/cashflow.pyx
	quantlib/currency.pyx
	quantlib/indexes/_euribor.pxd
	quantlib/indexes/_ibor_index.pxd
	quantlib/indexes/_interest_rate_index.pxd
	quantlib/indexes/_libor.pxd
	quantlib/indexes/_swap_index.pxd
	quantlib/indexes/euribor.pyx
	quantlib/indexes/ibor_index.pyx
	quantlib/indexes/interest_rate_index.pyx
	quantlib/indexes/libor.pyx
	quantlib/instruments/_bonds.pxd
	quantlib/instruments/_payoffs.pxd
	quantlib/instruments/bonds.pyx
	quantlib/pricingengines/_swap.pxd
	quantlib/pricingengines/swap.pyx
	quantlib/quotes.pyx
	quantlib/termstructures/_yield_term_structure.pxd
	quantlib/termstructures/yields/rate_helpers.pyx
	quantlib/termstructures/yields/yield_term_structure.pyx
	quantlib/test/test_calendar.py
	quantlib/test/test_date.py
	quantlib/time/_calendar.pxd
	quantlib/time/_daycounter.pxd
	quantlib/time/calendars/_jointcalendar.pxd
	quantlib/time/date.pyx
	quantlib/time/daycounter.pyx
	quantlib/time/daycounters/actual_actual.pyx
	quantlib/time/daycounters/thirty360.pyx
	setup.py

+++ quantlib/termstructures/volatility/equityfx/black_vol_term_structure.pyx
@@ -27,7 +27,7 @@ cdef class BlackConstantVol(BlackVolTermStructure):
     def __init__(self,
         Date reference_date,
         Calendar calendar,
-        float volatility,
+        double volatility,
         DayCounter daycounter
     ):
